refactor: ダッシュボードページのコンポーネント分割（564→228行）
9個のUIセクションを独立コンポーネントに抽出:
- OnboardingChecklist, SummaryCards, QuickActions
- InventoryAlert, TodayAppointments, BirthdayCustomers
- MonthlySales, RecentRecords, FeatureTip
