+++ src/components/dashboard/summary-cards.tsx
@@ -1,0 +1,67 @@
+import Link from "next/link";
+
+export function SummaryCards({
+  appointmentCount,
+  lapsedCount,
+  customerCount,
+}: {
+  appointmentCount: number;
+  lapsedCount: number;
+  customerCount: number;
+}) {
+  return (
+    <div className="grid grid-cols-2 gap-3">
+      <Link
+        href="/appointments"
+        className="bg-surface border border-border rounded-2xl p-4 hover:border-accent transition-colors"
+      >
+        <div className="flex items-center gap-3">
+          <div className="w-10 h-10 rounded-xl bg-accent/10 flex items-center justify-center shrink-0">
+            <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-5 h-5 text-accent">
+              <path strokeLinecap="round" strokeLinejoin="round" d="M6.75 3v2.25M17.25 3v2.25M3 18.75V7.5a2.25 2.25 0 0 1 2.25-2.25h13.5A2.25 2.25 0 0 1 21 7.5v11.25m-18 0A2.25 2.25 0 0 0 5.25 21h13.5A2.25 2.25 0 0 0 21 18.75m-18 0v-7.5A2.25 2.25 0 0 1 5.25 9h13.5A2.25 2.25 0 0 1 21 11.25v7.5" />
+            </svg>
+          </div>
+          <div>
+            <p className="text-2xl font-bold text-accent">{appointmentCount}</p>
+            <p className="text-xs text-text-light">今日の予約</p>
+          </div>
+        </div>
+      </Link>
+      {lapsedCount > 0 ? (
+        <Link
+          href="/customers"
+          className="bg-surface border border-orange-200 rounded-2xl p-4 hover:border-accent transition-colors"
+        >
+          <div className="flex items-center gap-3">
+            <div className="w-10 h-10 rounded-xl bg-warning/10 flex items-center justify-center shrink-0">
+              <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-5 h-5 text-warning">
+                <path strokeLinecap="round" strokeLinejoin="round" d="M12 9v3.75m9-.75a9 9 0 1 1-18 0 9 9 0 0 1 18 0Zm-9 3.75h.008v.008H12v-.008Z" />
+              </svg>
+            </div>
+            <div>
+              <p className="text-2xl font-bold text-warning">{lapsedCount}</p>
+              <p className="text-xs text-text-light">要フォロー</p>
+            </div>
+          </div>
+        </Link>
+      ) : (
+        <Link
+          href="/customers"
+          className="bg-surface border border-border rounded-2xl p-4 hover:border-accent transition-colors"
+        >
+          <div className="flex items-center gap-3">
+            <div className="w-10 h-10 rounded-xl bg-accent/10 flex items-center justify-center shrink-0">
+              <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-5 h-5 text-accent">
+                <path strokeLinecap="round" strokeLinejoin="round" d="M15 19.128a9.38 9.38 0 0 0 2.625.372 9.337 9.337 0 0 0 4.121-.952 4.125 4.125 0 0 0-7.533-2.493M15 19.128v-.003c0-1.113-.285-2.16-.786-3.07M15 19.128v.106A12.318 12.318 0 0 1 8.624 21c-2.331 0-4.512-.645-6.374-1.766l-.001-.109a6.375 6.375 0 0 1 11.964-3.07M12 6.375a3.375 3.375 0 1 1-6.75 0 3.375 3.375 0 0 1 6.75 0Zm8.25 2.25a2.625 2.625 0 1 1-5.25 0 2.625 2.625 0 0 1 5.25 0Z" />
+              </svg>
+            </div>
+            <div>
+              <p className="text-2xl font-bold">{customerCount}</p>
+              <p className="text-xs text-text-light">顧客数</p>
+            </div>
+          </div>
+        </Link>
+      )}
+    </div>
+  );
+}
